do some code refactoring
