+++ src/Mod/Mesh/App/Core/Projection.h
@@ -61,47 +61,6 @@ private:
     const MeshKernel& kernel;
 };
 
-#ifdef FC_USE_OCC
-/// Helper class
-struct SplitEdge
-{
-  unsigned long uE0, uE1; /**< start and endpoint of an edge */
-  Base::Vector3f cPt; /**< Point on edge (\a uE0, \a uE1) */
-};
-
-/**
- * The MeshProjection class projects a shape onto a mesh.
- * @author Werner Mayer
- */
-class MeshExport MeshProjection
-{
-public:
-  /// Construction
-  MeshProjection( const MeshKernel& rMesh);
-  /// Destruction
-  ~MeshProjection();
-
-  /**
-   * Searches all edges that intersect with the projected curve \a aShape. Therefore \a aShape must
-   * contain shapes of type TopoDS_Edge, other shape types are ignored. A possible solution is
-   * taken if the distance between the curve point and the projected point is <= \a fMaxDist.
-   */
-  void projectToMesh ( const TopoDS_Shape &aShape, float fMaxDist, std::vector<SplitEdge>& rSplitEdges ) const;
-  /**
-   * Cuts the mesh at the curve defined by \a aShape. This method call @ref projectToMesh() to get the
-   * split the facet at the found points. @see projectToMesh() for more details.
-   */
-  void splitMeshByShape ( const TopoDS_Shape &aShape, float fMaxDist ) const;
-
-protected:
-  void projectEdgeToEdge( const TopoDS_Edge &aCurve, float fMaxDist, const MeshFacetGrid& rGrid, 
-                          std::vector<SplitEdge>& rSplitEdges ) const;
-
-private:
-  const MeshKernel& _rcMesh;
-};
-#endif
-
 } // namespace MeshCore
 
 #endif  // MESH_PROJECTION_H
